Add support to provide digest algorithm when signing jar files
This is very useful when you have already signed jar file and when you
try to sign that file again on JDK 1.7+.

By default JDK 1.7+ use SHA256 for digest but older JDK use SHA-1. There
is a known bug in Jar sign verification that in this case will fail.
https://bugs.openjdk.java.net/browse/JDK-8079815

in this case you can specify jarsign to use sha-1 digest and there won't
be any problems with verification.